Remove deploy module and rename client module to microservices
